fastapi-leave-service/requirements.txt
# filepath: fastapi-leave-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic[email]==2.8.0
python-jose==3.3.0
firebase-admin==6.4.0
google-cloud-pubsub==2.30.0  # opcional, solo si vas a publicar en Pub/Sub
httpx==0.24.1